Added error handling and dto validation

+++ src/books/books.service.ts
@@ -1,40 +1,45 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { Book } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { CreateBookDto, UpdateBookDto } from './books.dto';
 
 @Injectable()
 export class BooksService {
   constructor(private readonly prisma: PrismaService) {}
 
-  getAll(): Promise<Book[]> {
-    return this.prisma.book.findMany();
+  async getAll(): Promise<Book[]> {
+    return await this.prisma.book.findMany();
   }
 
-  create(bookDto: {
-    title: string;
-    author: string;
-    year: number;
-    genre: string;
-  }): Promise<Book> {
-    return this.prisma.book.create({ data: bookDto });
+  async create(bookDto: CreateBookDto): Promise<Book> {
+    return await this.prisma.book.create({ data: bookDto });
   }
 
-  update(
-    id: string,
-    bookDto: {
-      title?: string;
-      author?: string;
-      year?: number;
-      genre?: string;
-    },
-  ): Promise<Book> {
-    return this.prisma.book.update({ where: { id }, data: bookDto });
+  async getOne(id: string): Promise<Book> {
+    const book = await this.prisma.book.findUnique({ where: { id } });
+    if (!book) {
+      throw new NotFoundException(`Book with ID ${id} not found`);
+    }
+    return book;
   }
 
-  getOne(id: string): Promise<Book> {
-    return this.prisma.book.findFirst({ where: { id } });
+  async update(id: string, bookDto: UpdateBookDto): Promise<Book> {
+    const book = await this.getOne(id);
+    if (!book) {
+      throw new NotFoundException(`Book with ID ${id} not found`);
+    }
+    return await this.prisma.book.update({
+      where: { id },
+      data: bookDto,
+    });
   }
-  delete(id: string): Promise<Book> {
-    return this.prisma.book.delete({ where: { id } });
+
+  async delete(id: string): Promise<Book> {
+    const book = await this.getOne(id);
+
+    if (!book) {
+      throw new NotFoundException(`Book with ID ${id} not found`);
+    }
+    return await this.prisma.book.delete({ where: { id } });
   }
 }
